Fixed race condition; Added WaitGroup support; Improved tests
Fixed a race condition in Program which would allow a program to be
executed twice at the same time, which would cause an intermittant test
failure. Did this by replacing the running variable with an atomic bool.

Added the AtomicBOol class and tests
Converted program tests into a test suite for code reuse
Added support for using a WaitGroup to SectionRunner and Program. The
start methods of these classes now also take a WaitGroup that can be
used to wait until the inner goroutine finishes.

Also generally improved tests. Added a test timeout to the Makefile